# Fix emoji decoding error