--- a/.travis.yml
+++ b/.travis.yml
@@ -16,6 +16,7 @@ script:
   - mvn clean package -DskipTests=true -f test_hessian/pom.xml
   - mvn test -f test_hessian/pom.xml
   - mvn clean package -f test_dubbo/pom.xml
+  - git reset --hard
   - go fmt && [[ -z `git status -s` ]]
   - sh before_validate_license.sh
   - chmod u+x /tmp/tools/license/license-header-checker