--- a/string.go
+++ b/string.go
@@ -255,9 +255,11 @@ func (d *Decoder) getStringLength(tag byte) (int, error) {
 
 func (d *Decoder) decString(flag int32) (string, error) {
 	var (
-		tag  byte
-		last bool
-		s    string
+		tag       byte
+		charTotal int
+		last      bool
+		s         string
+		r         rune
 	)
 
 	if flag != TAG_READ {
@@ -312,18 +314,24 @@ func (d *Decoder) decString(flag int32) (string, error) {
 			last = true
 		}
 
-		chunkLen, err := d.getStringLength(tag)
+		l, err := d.getStringLength(tag)
 		if err != nil {
 			return s, perrors.WithStack(err)
 		}
-		bytesBuf := make([]byte, chunkLen<<2)
-		offset := 0
+		charTotal = l
+		charCount := 0
+
+		runeData := make([]rune, charTotal)
+		runeIndex := 0
+
+		byteCount := 0
+		byteLen := 0
+		charLen := 0
 
 		for {
-			if chunkLen <= 0 {
+			if charCount == charTotal {
 				if last {
-					b := bytesBuf[:offset]
-					return *(*string)(unsafe.Pointer(&b)), nil
+					return string(runeData[:runeIndex]), nil
 				}
 
 				b, _ := d.ReadByte()
@@ -338,190 +346,21 @@ func (d *Decoder) decString(flag int32) (string, error) {
 						last = true
 					}
 
-					chunkLen, err = d.getStringLength(b)
+					l, err := d.getStringLength(b)
 					if err != nil {
 						return s, perrors.WithStack(err)
 					}
-					remain, cap := len(bytesBuf)-offset, chunkLen<<2
-					if remain < cap {
-						grow := len(bytesBuf) + cap
-						bs := make([]byte, grow)
-						copy(bs, bytesBuf)
-						bytesBuf = bs
-					}
+					charTotal += l
+					bs := make([]rune, charTotal)
+					copy(bs, runeData)
+					runeData = bs
+
 				default:
 					return s, perrors.New("expect string tag")
 				}
 			}
 
-			if chunkLen > 0 {
-				nread, err := d.next(bytesBuf[offset : offset+chunkLen])
-				if err != nil {
-					if err == io.EOF {
-						break
-					}
-					return s, perrors.WithStack(err)
-				}
-
-				// quickly detect the actual number of bytes
-				prev, i := offset, offset
-				len := offset + nread
-				copied := false
-				for r, r1 := len-1, len-2; i < len; chunkLen-- {
-					ch := bytesBuf[offset]
-					if ch < 0x80 {
-						i++
-						offset++
-					} else if (ch & 0xe0) == 0xc0 {
-						i += 2
-						offset += 2
-					} else if (ch & 0xf0) == 0xe0 {
-						// handle the 3-byte right edge
-						// case:
-						// 1. Expect 3 bytes, but the current byte is on the right
-						// 2. Expect 3 bytes, but the current byte is second to last to the right
-						if i == r {
-							bytesBuf[i+1], err = d.reader.ReadByte()
-							if err != nil {
-								return s, perrors.WithStack(err)
-							}
-							bytesBuf[i+2], err = d.reader.ReadByte()
-							if err != nil {
-								return s, perrors.WithStack(err)
-							}
-							nread += 2
-							len += 2
-						} else if i == r1 {
-							bytesBuf[i+2], err = d.reader.ReadByte()
-							if err != nil {
-								return s, perrors.WithStack(err)
-							}
-							nread++
-							len++
-						}
-
-						// we detect emoji first
-						c1 := ((uint32(ch) & 0x0f) << 12) + ((uint32(bytesBuf[i+1]) & 0x3f) << 6) + (uint32(bytesBuf[i+2]) & 0x3f)
-						if c1 >= 0xD800 && c1 <= 0xDBFF {
-
-							var (
-								c2  rune
-								n2  int
-								err error
-								ch0 byte
-							)
-
-							// more cache byte available
-							if i+3 < len {
-								ch0 = bytesBuf[i+3]
-							} else {
-								ch0, err = d.reader.ReadByte()
-								if err != nil {
-									return s, perrors.WithStack(err)
-								}
-								// update accumulates read bytes,
-								// because it reads more than thunk bytes
-								nread++
-								len++
-							}
-
-							if ch0 < 0x80 {
-								c2, n2 = rune(ch0), 1
-							} else if (ch0 & 0xe0) == 0xc0 {
-								var ch1 byte
-								if i+4 < len {
-									ch1 = bytesBuf[i+4]
-								} else {
-									// out of the chunk byte data
-									bytesBuf[i+4], err = d.reader.ReadByte()
-									if err != nil {
-										return s, perrors.WithStack(err)
-									}
-									ch1 = bytesBuf[i+4]
-									nread++
-									len++
-								}
-								c2, n2 = rune(((uint32(ch0)&0x1f)<<6)+(uint32(ch1)&0x3f)), 2
-							} else if (ch0 & 0xf0) == 0xe0 {
-								var ch1, ch2 byte
-								if i+5 < len {
-									ch1 = bytesBuf[i+4]
-									ch2 = bytesBuf[i+5]
-								} else {
-									ch1, err = d.reader.ReadByte()
-									if err != nil {
-										return s, perrors.WithStack(err)
-									}
-									ch2, err = d.reader.ReadByte()
-									if err != nil {
-										return s, perrors.WithStack(err)
-									}
-									len += 2
-									nread += 2
-								}
-								c := ((uint32(ch0) & 0x0f) << 12) + ((uint32(ch1) & 0x3f) << 6) + (uint32(ch2) & 0x3f)
-								c2, n2 = rune(c), 3
-							}
-
-							c := rune(c1-0xD800)<<10 + (c2 - 0xDC00) + 0x10000
-							n3 := utf8.EncodeRune(bytesBuf[i:], c)
-							if copied = n3 > 0 && n3 < /** front three byte */ 3+n2; copied {
-								// We need to move the bytes,
-								// for example, less bytes after decoding
-								offset = i + n3
-								copy(bytesBuf[offset:], bytesBuf[i+3+n2:len])
-							}
-
-							i += n2
-							chunkLen--
-						}
-						i += 3
-
-						// fix read the next byte index
-						if copied {
-							copied = false
-							continue
-						}
-
-						offset += 3
-					} else {
-						return s, perrors.Errorf("bad utf-8 encoding")
-					}
-				}
-
-				if remain := offset - prev - nread; remain > 0 {
-					if remain == 1 {
-						ch, err := d.ReadByte()
-						if err != nil {
-							return s, perrors.WithStack(err)
-						}
-						bytesBuf[offset-1] = ch
-					} else {
-						var err error
-						if buffed := d.Buffered(); buffed < remain {
-							// trigger fill data if required
-							copy(bytesBuf[offset-remain:offset], d.peek(remain))
-							_, err = d.reader.Discard(remain)
-						} else {
-							// copy remaining bytes.
-							_, err = d.next(bytesBuf[offset-remain : offset])
-						}
-
-						if err != nil {
-							return s, perrors.WithStack(err)
-						}
-					}
-				}
-
-				// the expected length string has been processed.
-				if chunkLen <= 0 {
-					// we need to detect next chunk
-					continue
-				}
-			}
-
-			// decode byte
-			ch, err := d.ReadByte()
+			r, charLen, byteLen, err = decodeUcs4Rune(d.reader)
 			if err != nil {
 				if err == io.EOF {
 					break
@@ -529,58 +368,14 @@ func (d *Decoder) decString(flag int32) (string, error) {
 				return s, perrors.WithStack(err)
 			}
 
-			if ch < 0x80 {
-				bytesBuf[offset] = ch
-				offset++
-			} else if (ch & 0xe0) == 0xc0 {
-				ch1, err := d.ReadByte()
-				if err != nil {
-					return s, perrors.WithStack(err)
-				}
-				bytesBuf[offset] = ch
-				bytesBuf[offset+1] = ch1
-				offset += 2
-			} else if (ch & 0xf0) == 0xe0 {
-				var err error
-				if buffed := d.Buffered(); buffed < 2 {
-					// trigger fill data if required
-					copy(bytesBuf[offset+1:offset+3], d.peek(2))
-					_, err = d.reader.Discard(2)
-				} else {
-					_, err = d.next(bytesBuf[offset+1 : offset+3])
-				}
-				if err != nil {
-					return s, perrors.WithStack(err)
-				}
+			runeData[runeIndex] = r
+			runeIndex++
 
-				bytesBuf[offset] = ch
-
-				// we detect emoji first
-				c1 := ((uint32(ch) & 0x0f) << 12) + ((uint32(bytesBuf[offset+1]) & 0x3f) << 6) + (uint32(bytesBuf[offset+2]) & 0x3f)
-				if c1 >= 0xD800 && c1 <= 0xDBFF {
-					c2, n2, err := decodeUcs2Rune(d.reader)
-					if err != nil {
-						return s, perrors.WithStack(err)
-					}
-
-					c := rune(c1-0xD800)<<10 + (c2 - 0xDC00) + 0x10000
-					utf8.EncodeRune(bytesBuf[offset:], c)
-
-					// update next rune
-					offset += n2
-					chunkLen--
-				}
-
-				offset += 3
-			} else {
-				return s, perrors.Errorf("bad utf-8 encoding, offset=%d\n", offset)
-			}
-
-			chunkLen--
+			charCount += charLen
+			byteCount += byteLen
 		}
 
-		b := bytesBuf[:offset]
-		return *(*string)(unsafe.Pointer(&b)), nil
+		return string(runeData[:runeIndex]), nil
 	}
 
 	return s, perrors.Errorf("unknown string tag %#x\n", tag)