--- a/test_hessian/src/main/java/test/TestString.java
+++ b/test_hessian/src/main/java/test/TestString.java
@@ -30,6 +30,13 @@ public class TestString {
         return s + ",max" + maxUnicode;
     }
 
+    /**
+     * see https://github.com/apache/dubbo-go-hessian2/issues/252
+     */
+    public static String getEmojiTestString2() {
+        return "❄️\uD83D\uDEAB\uD83D\uDEAB\uD83D\uDEAB\uD83D\uDEAB 多次自我介绍、任务、动态和";
+    }
+
     public static String getComplexString() {
         String s = "킐\u0088中国你好!\u0088\u0088\u0088\u0088\u0088\u0088";
         return s;