<!--  Thanks for sending a pull request! 
-->

**What this PR does**:
fix emoji decoding error and refactor string decoding algorithm, with 17% performance improvement

**Which issue(s) this PR fixes**:

Fixes #252 #253

**Special notes for your reviewer**:

**Does this PR introduce a user-facing change?**:
```release-note
- fix emoji decoding error and refactor string decoding algorithm, with 17% performance improvement
```