@wongoo 2021-01-03
## benchmark result

before fixing:
```
BenchmarkDecodeStringAscii
BenchmarkDecodeStringAscii-8                6952            173795 ns/op
BenchmarkDecodeStringUnicode
BenchmarkDecodeStringUnicode-8              5228            229713 ns/op
BenchmarkDecodeStringEmoji
    string_test.go:230: err: bad utf-8 encoding
--- FAIL: BenchmarkDecodeStringEmoji
```

after fixing:
```
BenchmarkDecodeStringAscii
BenchmarkDecodeStringAscii-8                7149            145317 ns/op
BenchmarkDecodeStringUnicode
BenchmarkDecodeStringUnicode-8              5768            182145 ns/op
BenchmarkDecodeStringEmoji
BenchmarkDecodeStringEmoji-8                1400            816080 ns/op
```

with 17% performance improvement.