@fangyincheng 2021-01-05
ci failed 
```The command "/tmp/tools/license/license-header-checker -v -a -r -i vendor  /tmp/tools/license/license.txt . go  && [[ -z `git status -s` ]]" exited with 1.```